fixed ip field in otp table

+++ app/Containers/Authorization/Data/Migrations/2020_03_19_103618_create_otp_tokens_table.php
@@ -22,7 +22,7 @@ class CreateOtpTokensTable extends Migration
             $table->string('reason', 20)->nullable();
             $table->boolean('used')->default(false);
             $table->unsignedInteger('ttl')->default(300)->comment('in seconds');
-            $table->unsignedInteger('ip')->comment('requester ip');
+            $table->unsignedBigInteger('ip')->comment('requester ip');
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('restrict');
